Remove typecasts and place loops in appropriate files
