+++ bot/src/auctions.ts
@@ -1,6 +1,8 @@
 import type { AuctionInitialInfo } from 'auctions-core/src/types';
 import { fetchAllInitialAuctions } from 'auctions-core/src/auctions';
 import { getWhitelistedCollaterals } from './whitelist';
+import { notifyCollateral } from './notify';
+import participate from './keeper';
 
 const THRESHOLD_FOR_NEW_AUCTIONS = 5 * 60 * 1000;
 const knownAuctionIds = new Set();
@@ -31,4 +33,18 @@ export const getAllAuctions = async function (network: string): Promise<AuctionI
     const auctionIds = auctions.map(auction => `"${auction.id}"`).join(', ');
     console.info(`auctions: found "${auctions.length}" auctions (${auctionIds}) on "${network}" network`);
     return auctions;
+};
+
+export const loopCollateral = async function (network: string): Promise<void> {
+    try {
+        const activeAuctions = await getAllAuctions(network);
+        if (activeAuctions.length === 0) {
+            return;
+        }
+        const newAuctions = getNewAuctionsFromActiveAuctions(activeAuctions);
+        newAuctions.map(notifyCollateral);
+        participate(network, activeAuctions);
+    } catch (error) {
+        console.error('loop error:', error);
+    }
 };
